# docs/design/headless-ravn-v1.openapi.yaml
openapi: 3.1.0
info:
  title: RAVN Headless Broker API
  version: 0.3.0-design
  summary: Proposed MCP-first broker; REST management and alternative calls, with optional policy extensions.
  description: |
    DESIGN CONTRACT ONLY. This specification is a proposal for the headless RAVN
    v1 implementation. It does not describe a shipping or integration-tested API.
    The proposed server is Python/FastAPI with the official Python MCP SDK,
    one Uvicorn worker, and SQLite; this HTTP contract is language-independent.

    MCP-FIRST DELIVERY. REST manages account onboarding, connection lifecycle,
    runtime-session issuance/revocation, and core metadata-only activity listing. Agents primarily
    discover and call tools through /mcp using connection-bound session tokens.
    REST discovery/execution remains a supported backend alternative through the
    SAME authorization/execution engine; developers need not integrate both call
    paths. Connecting an account is setup, not something an agent repeats per
    call. Runtime-token expiry does not disconnect the provider account.
    Personal bearer import supports development and intentional bring-your-own-
    token deployments; GitHub App user OAuth is the default onboarding profile.
    Fine-grained policy PATCH and events queries are OPTIONAL EXTENSIONS.
    No permanent agent registration or duplicate permission editor is required.

    RAVN authenticates an application credential and derives app_id from that
...

...

...

...

...
servers:
...
security:
  - ApplicationCredential: []
...
paths:
...
  /v1/connect-sessions:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/UserId'
    post:
      operationId: startConnectSession
      tags: [Connect sessions]
      summary: Start a user-owned OAuth connection or reconnect transaction.
      description: |
        The authenticated actor becomes the connection owner. The integration
        must be enabled for the application and support OAuth. return_url must
        exactly match an operator-approved application return URL; wildcard,
        suffix, or attacker-controlled redirects are not accepted.

        app_state is an opaque application-generated transaction value. The
        application stores it against the initiating user's browser/login
        transaction and verifies it before completing the flow. Do not encode
        personal data or secrets in it. RAVN independently manages provider OAuth
        state and PKCE where supported/required; app_state does not replace them.

        The whole authorization transaction expires after 10 minutes. After a
        validated provider callback, credentials are staged until the earlier
        of callback time plus 5 minutes and the original transaction expiry;
        staging never extends the original 10-minute lifetime. A single-use
        completion_code is delivered only through the
        browser return to the approved return_url, together with app_state and
        the connect-session identifier. Credentials are not committed until the
        application backend calls complete. GET status never returns that code.

        For reconnect, the existing connection must belong to this actor and use
        this integration. Its previously verified provider account/workspace
        identity must match the new authorization. A different account requires
        a new connection. A failed or cancelled attempt leaves the old connection
        and its credentials unchanged. OAuth grants remain provider-controlled;
        app-defined RAVN restrictions are configured separately.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartConnectSessionRequest'
            examples:
              newConnection:
                value:
                  integration_id: github
                  return_url: https://app.example.com/settings/connections/return
                  app_state: xQM2Tt90y7iueS9svXq9I2UluMX8ev11
              reconnect:
                value:
                  integration_id: github
                  return_url: https://app.example.com/settings/connections/return
                  app_state: eDBvfLxLBcAKeqTrDoznCEsAZcv9AH3M
                  reconnect_connection_id: conn_01K4ZAZBR1KQEM19YEJGJ3RF5P
      responses:
        '201':
          description: Pending OAuth transaction created.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
            Location:
              description: Actor-authorized status endpoint for the new transaction.
              schema:
                type: string
              example: /v1/connect-sessions/cs_01K4ZAZBR1KQEM19YEJGJ3RF5P
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedConnectSession'
              example:
                id: cs_01K4ZAZBR1KQEM19YEJGJ3RF5P
                status: pending
                authorization_url: https://ravn.example.com/connect/cs_01K4ZAZBR1KQEM19YEJGJ3RF5P/start
                expires_at: '2026-09-11T20:10:00Z'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/Error'
  /v1/connect-sessions/{connect_session_id}:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/UserId'
      - $ref: '#/components/parameters/ConnectSessionId'
    get:
      operationId: getConnectSession
      tags: [Connect sessions]
      summary: Read onboarding status without receiving a completion code or credentials.
      description: |
        Only the originating application/tenant/user context can read the
        transaction. This endpoint does not return authorization_url,
        completion_code, provider authorization codes, tokens, or OAuth state.
        connection_id is populated only after successful completion.
      responses:
        '200':
          description: Current sanitized transaction status.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectSession'
              example:
                id: cs_01K4ZAZBR1KQEM19YEJGJ3RF5P
                integration_id: github
                status: awaiting_completion
                expires_at: '2026-09-11T20:10:00Z'
                completion_expires_at: '2026-09-11T20:08:00Z'
                connection_id: null
                created_at: '2026-09-11T20:00:00Z'
                updated_at: '2026-09-11T20:03:00Z'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /v1/connect-sessions/{connect_session_id}/complete:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/UserId'
      - $ref: '#/components/parameters/ConnectSessionId'
    post:
      operationId: completeConnectSession
      tags: [Connect sessions]
      summary: Commit staged credentials after verifying the original application transaction.
      description: |
        Before calling this endpoint, the application must match the returned
        connect-session ID and app_state to its original server-side browser
        transaction and authenticated user. Merely accepting a flow identifier
        submitted by any logged-in browser is not sufficient.

        completion_code arrives only through the browser return. It is
        short-lived, single-use, bound to this transaction and actor, and stored
        only as a verifier/hash by RAVN. Validate and consume it atomically with
        the connection commit. Replaying it returns completion_code_invalid.
        If the application loses a successful response, it can read the status
        endpoint to recover the committed connection_id without reusing the code.

        Same-account reconnect atomically replaces the active credential and
        invalidates older runtime sessions by incrementing connection epoch.
        Pending credentials and codes are destroyed after expiry/cancellation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteConnectSessionRequest'
      responses:
        '200':
          description: Connection committed; no credentials are returned.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
            ETag:
              $ref: '#/components/headers/ConnectionETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Connection'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/Error'
...
  /v1/connections/{connection_id}/disconnect:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/UserId'
      - $ref: '#/components/parameters/ConnectionId'
    post:
      operationId: disconnectConnection
      tags: [Connections]
      summary: Disable a personal connection and invalidate its runtime sessions.
      description: |
        Commit local disabled state and increment epoch before acknowledging.
        Every subsequent call checks that state before dispatch. This operation
        is locally idempotent and preserves an audit tombstone; it is not a
        destructive metadata-delete endpoint.

        Already-dispatched upstream operations may finish and are not undone.
        Provider-side revocation is a distinct best-effort action, depending on
        provider capability. Never claim global provider revocation when only
        local disconnect succeeded. API callers cannot disconnect an
        operator-managed service connection merely because they may execute it.
      responses:
        '200':
          description: Local disconnect committed, with separate provider revocation status.
          headers:
            ETag:
              $ref: '#/components/headers/ConnectionETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DisconnectResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
...
  /v1/sessions/{session_id}/revoke:
    parameters:
      - $ref: '#/components/parameters/TenantId'
      - $ref: '#/components/parameters/UserId'
      - $ref: '#/components/parameters/SessionId'
    post:
      operationId: revokeRuntimeSession
      x-ravn-phase: core
      tags: [Runtime sessions]
      summary: Revoke a runtime credential without disconnecting its provider account.
      description: |
        Locally idempotent. After the state commit, new dispatches authenticated
        with this token fail even before its original expiry. This does not
        revoke the application's credential, the provider credential, other
        sessions, or already-dispatched upstream work.
      responses:
        '204':
          description: Revoked, or already revoked/expired.
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    ApplicationCredential:
      type: http
      scheme: bearer
      bearerFormat: Opaque RAVN application credential
      description: |
        A secret provisioned for one configured application. RAVN derives app_id
        from it. Keep it in the trusted backend and never in browser bundles,
        mobile applications, model prompts, or an agent sandbox. It is distinct
        from provider OAuth credentials and from connection-bound runtime tokens.
        It can assert any user in its application's namespace; it does not grant
        RAVN deployment administration or identity for a particular agent.
  parameters:
    TenantId:
      name: X-Ravn-Tenant-Id
      in: header
      required: false
      description: |
        CONDITIONALLY REQUIRED by operator-configured tenant_mode, not by caller
        choice. In multi mode, omission or an empty value returns invalid_request.
        In single mode, omission resolves to the literal default; an explicit
        default is accepted and any other value returns invalid_request.
        OpenAPI cannot express a header requirement conditional on server-side
        application configuration. Every resource is still scoped by the resolved
        tenant. Derive this identifier from verified application context, never
        model output. Do not concatenate tenant and user IDs to emulate isolation.
      schema:
        $ref: '#/components/schemas/ActorIdentifier'
      example: tenant_acme
...
    ConnectSessionId:
      name: connect_session_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/ResourceId'
...
  headers:
    NoStore:
      description: Sensitive response is not cacheable.
      schema:
        type: string
        const: no-store
...